% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_netp.R
\name{compute_netp}
\alias{compute_netp}
\title{This script will compute the net attentive engagement metric given a dataset}
\usage{
compute_netp(combined.data)
}
\description{
Reaction types are collected at the daily but not gender/age level
so the denominator there is larger
}
\details{
First, turn each of the metrics into a proportion of Reach

We used a glm version of a weighted average
I have used IRT and other approaches, but this is simplest and seems most stable
In the past, I have also weighted this by Reach, but that tends to create biases
when FB has been allowed to optimize. In this case, we have a couple of lines where Denominator is
many many thousands, so taking the weights off shows the real patterns more clearly

Likes are added back

The Response row sum is a rough proportion of all Denominator that had any of the responses
there could be duplication, but, in practice, it's not a big deal

This is net Attentive engagement. It's Attentive minus not Attentive.
because of the larger than 1 coefficients, it is possible for Attentive engagement to be greater than the total responses
In that case, we assume that negative was zero

data will need the following columns
'angry','haha','like','love','sad','wow'
'Post.Comments', 'Post.Shares'
Ad.Name,
Denominator
}
