% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_var.R
\name{extract_var}
\alias{extract_var}
\title{Extract Text from Variable of Interest}
\usage{
extract_var(
  data,
  from = "survey",
  to = "var",
  keys = NA,
  labels = NA,
  remove.dash = TRUE,
  remove.underscore = TRUE,
  capitalize = TRUE
)
}
\arguments{
\item{data}{Data containing variables of interest}

\item{from}{The varible from which text will be extracted}

\item{to}{The variable to be created containing values specified in "keys"}

\item{keys}{List of strings to be extracted from the variable specified in "from"}

\item{labels}{List of labels to be use instead of raw values in "keys"}

\item{remove.dash}{Whether to remove dash symbols in values of the variable specifed in "to" (Default to 'TRUE')}

\item{remove.underscore}{Whether to remove underscore symbols in values of the variable specifed in "to" (Default to 'TRUE')}

\item{capitalize}{Whether to capitalized values in the variable specifed in "to" (Default to 'TRUE')}
}
\value{
Original dataset with an additional column containing the variable named in "to" with values in "keys" extracted from the variable specified in "from"
}
\description{
This functions allows for the extraction of text from a variable to make a new variable.
}
\details{
Developer note: Very useful when use with read_all()
}
\examples{
dat <- df \%>\% extract_var(from = "survey", to = "parent", keys = c("parents", "non-parents"), labels = c("With Children","Without Children"))

}
