% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_fb_data.R
\name{format_fb_data}
\alias{format_fb_data}
\title{Format data from FB}
\usage{
format_fb_data(df_Reactions, df_Reach)
}
\description{
df_Reactions must have the following columns
Post Reaction Type
Post Reactions
Ad Name
}
\details{
df_Reach must have the following column(s)
Ad Name
}
