% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_power1.R
\name{get_power1}
\alias{get_power1}
\title{Compute GPA's Persuasion Test Power Stasitics}
\usage{
get_power1(
  data,
  treatment,
  suvey_q,
  prop = "percent",
  size = "total",
  alt = "greater"
)
}
\arguments{
\item{data}{Data containing survey response data (pre-processed by hypos_prep)}

\item{treatment}{Column specifying treatments of interest (e.g., ad_name)}

\item{size}{Column containing sample size for each treatment (default to be "total")}

\item{alt}{Alternative hypothesis; can be "greater", "less", or "two.sided" (default to be "greater")}

\item{survey_q}{The survey question of interest to summarise by}

\item{counts}{Column containing response counts for the question of interest (default to be "responses")}
}
\value{
A new data frame that contains an additional column for power statistics
}
\description{
This function generates the power values associated with
several treatments against a control, for a single survey question where
SAMPLE SIZE IS EQUAL. For unequal sample size, see get_power2
}
\examples{
get_power1(data = df, treatment = "ads", suvey_q = "q4",  counts="responses", size = "total", alt="greater")
}
