% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_power2.R
\name{get_power2}
\alias{get_power2}
\title{Compute GPA's Persuasion Test Power Stasitics}
\usage{
get_power2(
  data,
  treatment,
  survey_q,
  prop = "percent",
  sample1 = value1,
  sample2 = value2,
  alt = "greater"
)
}
\arguments{
\item{data}{Data containing survey response data (pre-processed by hypos_prep or equivalent)}

\item{treatment}{Column specifying treatments of interest (e.g., ad_name)}

\item{survey_q}{The survey question of interest}

\item{prop}{Column containing proportions for the response of interest (default to be "responses")}

\item{sample1}{Sample size for treatments}

\item{sample2}{Sample size for control}

\item{alt}{Alternative hypothesis; can be "greater", "less", or "two.sided" (default to be "greater")}
}
\value{
A new data frame containing input data plus an additional column for power values
}
\description{
This function generates the power values associated with
several treatments against a control, for a single survey question, where
the sample size is different between treatment and control.
}
\examples{
get_power1(data = df, treatment = "ads", suvey_q = "Q4",  n1=1200, n2=600, alt="greater")
}
