% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggsort.R
\name{ggsort}
\alias{ggsort}
\title{Sort Variables by Attributes}
\usage{
ggsort(data, var, by, attr = NA, ctrl_posn = "Default", decreasing = FALSE)
}
\arguments{
\item{data}{The data frame that contains the variables of interest}

\item{var}{The variable to be sorted}

\item{by}{Quanitative variable that determine the order}

\item{attr}{Attribute variable to filter (Default to NA)}

\item{ctrl_posn}{Position of the control variable: "Top", "Bottom", "Default"}

\item{decreasing}{Decreasing order of the variable specified in "by" (Default to "decreasing")}
}
\value{
Data frame with "var" sorted by "by" and attributes specified in "attr" and the position of the control specified in "ctrl_posn"
}
\description{
This function allows users to sort by two variables with specified attributes
and position the control variable at a specified location
}
\details{
Developer notes: This function works exceptionally well when you need to prepare data for ggplot.As example,
users need to sort the ad variable by increasing percentage but only when the answer for question 5 is "Trustworthy".
}
\examples{
ad_analysis <- ad_analysis \%>\% ggsort(var = "text", by = "netp", attr = c("age", "Over 35"), ctrl_posn = "Bottom")
}
