% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hypos_prep.R
\name{hypos_prep}
\alias{hypos_prep}
\title{Data Preparation for P-value and Power Statistics Computation}
\usage{
hypos_prep(data, treatment, survey_q, q_value)
}
\arguments{
\item{data}{Data containing survey response data}

\item{treatment}{The treatment of interest (e.g., ad_name, theme, etc.)}

\item{survey_q}{The survey question of interest to summarise by}

\item{q_value}{The value of interest to filter by.}
}
\value{
A smaller data frame ready for analysis
}
\description{
This function summarises the data by specific question(s) and its value in the survey data.
}
\examples{
hypos_prep(data = survey, treatment = "ad", survey_q = "q4", value = "agree")
}
