% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_all.R
\name{read_all}
\alias{read_all}
\title{Read All Data in the Directory}
\usage{
read_all(dir, sheet = "Individuals Coded")
}
\arguments{
\item{dir}{The directory where all of your data are located}

\item{sheet}{The name of the excel sheet where your data are located}
}
\value{
A combined data frame of all data in the directory or a list of combined like-datasets
}
\description{
This function allows user to read and combine all like-data with the extension .csv, .xlsx, or .rds in a folder.
}
\details{
Developer note: read_all() is essectially useful when reading Pollfish data for a persuasion test.
}
\examples{
df<- read_all("C:/Users/", "Individuals Coded")

}
