% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_fb.R
\name{read_fb}
\alias{read_fb}
\title{Read and Format GPA Facebook Data}
\usage{
read_fb(df_reactions_dir, df_engagement_dir)
}
\arguments{
\item{df_reactions_dir}{Directory to GPA Facebook post reaction types data (must contains 'post_reaction_type', 'post_reactions', and 'ad_name')}

\item{df_engagement_dir}{Directory to GPA Facebook post engagement data (must contains 'ad_name')}
}
\value{
A combined dataset with additional columns 'audience' and 'ad_id' needed for the next step of analysis.
}
\description{
This function will read in two data frames, i.e., the post reaction types data and the post engagement data, combine and ready them for analysis.
}
\examples{
read_fb(df_reactions_dir = "reaction_data.csv", df_engagement_dir = "engagement_data.csv")
}
