% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seg_prep.R
\name{seg_prep}
\alias{seg_prep}
\title{Re-format GPA for Audience Segmentation}
\usage{
seg_prep(data, start_idx)
}
\arguments{
\item{data}{Data containing survey response data}

\item{start_idx}{The column index that correspond to the first value question in the survey}
}
\value{
A new data frame containing input data plus an additional column for power values
}
\description{
This function gets the survey data ready for audience segmentation model
}
\examples{
get_seg(data = df, start_idx= 35)
}
