% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simple_cap.R
\name{simple_cap}
\alias{simple_cap}
\title{Capitalize the first letter of every word}
\usage{
simple_cap(y)
}
\arguments{
\item{x}{A string object or vector}
}
\value{
A captitalized string or vector
}
\description{
This function will capitalize the first letter of every word in a given string
}
\examples{
ad.info$text <- simple_cap(ad.info$text)
}
