% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summ_netp.R
\name{summ_netp}
\alias{summ_netp}
\title{Summazise Net Attentive Engagement by Specified Variable(s)}
\usage{
summ_netp(data, var1, var2, var3)
}
\arguments{
\item{data}{Dataset containing engagement metrics at the ad level (must contain attentive_engagement, denominator, responses)}

\item{var1}{- first variable to group by}

\item{var2}{- second variable to group by (optional)}

\item{var3}{- third variable to group by (optional)}
}
\value{
A smaller dataframe that contains the weighted mean of net attentive engagement by the specified variable(s)
}
\description{
This function groups engagement test data along up to three variables of interest,
and computes the denominator, positive engagement, negative engagement,
and net attentive engagement for each resulting group. It is most useful before plotting.
}
\examples{
summarize_netp(data = df, var1 = ad_name, var2 = age)
}
