% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_engagement.R
\name{summarize_engagement}
\alias{summarize_engagement}
\title{The function below groups engagement test data along up to three variables of interest,
and computes the reach, positive engagement, negative engagement,
and net positive engagement for each resulting group. It is most useful before plotting.}
\usage{
summarize_engagement(
  input_df,
  eng_var,
  resp_var,
  weight_var,
  group_var1,
  group_var2,
  group_var3
)
}
\description{
Required Parameters:
}
\details{
input_df - dataset containing engagement metrics at the ad level
eng_var - typically corresponds to "Attentive.engagement"
resp_var -  typically corresponds to "responses"
weight_var - typically corresponds to "Denominator"
group_var1 - first variable to group by (the other two are optional)
}
