% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrap_labels.R
\name{wrap_labels}
\alias{wrap_labels}
\title{Label Text Wrapping}
\usage{
wrap_labels(x, len)
}
\arguments{
\item{x}{Vector of strings or string object}

\item{len}{Maximum length of string per line}
}
\value{
Wrapped vector of strings or wrapped string object
}
\description{
This function wraps texts in the specified column by the number of character specified without 
breaking words.
}
\examples{
wrap_labels(df$labels, 50)
}
