% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_alpha.R
\name{get_alpha}
\alias{get_alpha}
\title{This function generates the p-values associated with
several treatments against a SINGLE control, for a single survey question.}
\usage{
get_alpha(df, treatment, counts, size, alt)
}
\description{
#Preprocessing that generates input DF usually looks like this:
Q4.ad.analysis <- combined.data %>% group_by(Ad,Q4) %>% summarise(n_q4=n()) %>% group_by(Ad) %>%
mutate(total_q4 = sum(n_q4)) %>% mutate(percent_q4 = n_q4/total_q4) %>% filter(Q4 == 1)
}
\details{
Required Parameters:

df - df containing survey response data, summarized at the treatment level and containing counts of responses for
several treatments against a SINGLE control, for a single survey question.

Required Parameters:

df - df containing survey response data
treatment - column specifying treatments of interest
counts - column containing response counts
size - column containing sample size for each treatment
alt - alternative hypothesis; can be "greater", "less", or "two.sided"
NOTE: COLUMN NAMES MUST BE PASSED AS STRINGS
}
