% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_alpha.R
\name{get_alpha}
\alias{get_alpha}
\title{This function generates the p-values associated with
several treatments against a control, for a single survey question.}
\usage{
get_alpha(df, treatment, counts, size, alt)
}
\description{
Required Parameters:
}
\details{
df - df containing survey response data
treatment - column specifying treatments of interest
counts - column containing response counts 
size - column containing sample size for each treatment 
alt - alternative hypothesis; can be "greater", "less", or "two.sided"
NOTE: COLUMN NAMES MUST BE PASSED AS STRINGS
}
