% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_alpha.R
\name{get_alpha}
\alias{get_alpha}
\title{Compute GPA's Persuasion Test P-values}
\usage{
get_alpha(
  data,
  treatment,
  survey_q,
  counts = "responses",
  size = "total",
  alt = "greater",
  control = "Control"
)
}
\arguments{
\item{data}{Data containing survey response data (pre-processed by hypos_prep or equivalent)}

\item{treatment}{Column specifying treatments of interest (e.g., ad_name)}

\item{survey_q}{The survey question of interest to summarise by}

\item{counts}{Column containing response counts for the question of interest (default to be "responses")}

\item{size}{Column containing sample size for each treatment (default to be "total")}

\item{alt}{Alternative hypothesis; can be "greater", "less", or "two.sided" (default to be "greater")}

\item{control}{Variable serving as the control (default to "Control")}
}
\value{
A new data frame containing input data plus an additional column for p-values
}
\description{
This function generates the p-values associated with
several treatments of inteterest against a single control, for responses to a single survey question.
}
\details{
Data must have been preprocessed by hypos_prep() or equivalent, and should contain survey response data
summarized at the desired treatment level, for a single survey question. Preprocessing that generates input data usually looks like this:
input_df <- survey_data %>% group_by(treatment,question) %>% summarise(responses=n()) %>% group_by(treatment) %>%
mutate(total = sum(responses)) %>% mutate(percent = responses/total) %>% filter(question == value)

Developer note: users can take advantages of the "control" parameter to  perform hypothesis testing in the event no control variable is present.
}
\examples{
get_alpha(data = df, treatment = "ads", suvey_q = "q4", counts="responses", size = "total", alt="greater", control = "Control")
get_alpha(data = df, treatment = "ads", suvey_q = "q4", counts="responses", size = "total", alt="greater", control = "Biden")
}
