% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_power1.R
\name{get_power1}
\alias{get_power1}
\title{Compute GPA's Persuasion Test Power Stasitics}
\usage{
get_power1(
  data,
  treatment,
  survey_q,
  prop = "percent",
  size = "total",
  alt = "one.sided"
)
}
\arguments{
\item{data}{Data containing survey response data (pre-processed by hypos_prep or equivalent)}

\item{treatment}{Column specifying treatments of interest (e.g., ad_name)}

\item{survey_q}{The survey question of interest}

\item{size}{Column containing sample size for each treatment (default to be "total")}

\item{alt}{Alternative hypothesis; can be "one.sided" or "two.sided" (default to be "one.sided")}

\item{counts}{Column containing response counts for the response of interest (default to be "responses")}
}
\value{
A new data frame containing input data plus an additional column for power values
}
\description{
This function generates the power values associated with
several treatments against a control, for a single survey question where
the sample size is equal across all (treatment and control). For unequal sample size, see get_power2()
}
\examples{
get_power1(data = df, treatment = "ads", suvey_q = "q4",  counts="responses", size = "total", alt="greater")
}
