% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_power1.R
\name{get_power1}
\alias{get_power1}
\title{This function generates the power values associated with
several treatments against a control, for a single survey question where 
SAMPLE SIZE IS EQUAL. FOR UNEQUAL SAMPLE SIZES, SEE get_power2}
\usage{
get_power1(df, treatment, prop, size, alt)
}
\description{
Required Parameters:
}
\details{
df - df containing survey response data
treatment - column specifying treatments of interest
prop - column containing response proportions
size - sample size; must be numeric
alt - alternative hypothesis; "one.sided" or "two.sided"
NOTE: COLUMN NAMES MUST BE PASSED AS STRINGS
}
