% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hypothesis_prep.R
\name{hypos_prep}
\alias{hypos_prep}
\title{Data Preparation for P-value and Power Statistics Computation}
\usage{
hypos_prep(data, treatment, survey_q, value)
}
\arguments{
\item{data}{Data containing survey response data}

\item{treatment}{The treatment of interest (e.g., ad_name, theme, etc.)}

\item{survey_q}{The survey question of interest to summarise by}

\item{value}{The value of interest to filter by.}
}
\value{
A smaller data frame that contains place-holder columns for p-value and power statsitics
}
\description{
This function summarises the data by specific question(s) and its value in the survey data.
}
\examples{
hypos_prep(data = survey, treatment = ad, survey_q = q4, value = "agree")
}
