% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hypos_prep.R
\name{hypos_prep}
\alias{hypos_prep}
\title{Data Preparation for P-value and Power Statistics Computation}
\usage{
hypos_prep(data, treatment, survey_q, q_value)
}
\arguments{
\item{data}{Data containing survey response data}

\item{treatment}{Column containing the treatments of interest (e.g., ad_name, theme, etc.)}

\item{survey_q}{The survey question of interest to summarise by}

\item{q_value}{The response of interest to filter by}
}
\value{
A smaller data frame ready for analysis
}
\description{
This function summarises survey reponse data at the desired treatment level, 
for a single question and a single response of interest.
}
\examples{
hypos_prep(data = survey, treatment = "ad", survey_q = "q4", value = "agree")
}
